Isolate more GC-specific data to `store/gc.rs`
This is an attempt to move more information that's only needed for
GC-enabled Wasmtime to `gc`-feature-gated files. This moves some
methods, types, etc, from `store.rs` to `store/gc.rs`
